# models/buildrequest.go
package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/validate"
)

/*Buildrequest buildrequest

swagger:model buildrequest
*/
type Buildrequest struct {

	/* git branch checkout
	 */
	Branch string `json:"branch,omitempty"`

	/* git commit to checkout
	 */
	Commit string `json:"commit,omitempty"`

	/* url to the git repo containing a packer config

	Required: true
	*/
	Giturl *string `json:"giturl"`

	/* path within the giturl repo to the packer config.  defaults to /packer.json
	 */
	Templatepath string `json:"templatepath,omitempty"`
}

// Validate validates this buildrequest
func (m *Buildrequest) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateGiturl(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Buildrequest) validateGiturl(formats strfmt.Registry) error {

	if err := validate.Required("giturl", "body", m.Giturl); err != nil {
		return err
	}

	return nil
}